Includes: Refactor: Move CWalletInterface and CMainSignals out of main
