+++ src/makefile.linux-mingw
@@ -65,6 +65,7 @@ HEADERS = $(wildcard *.h)
 OBJS= \
     obj/alert.o \
     obj/allocators.o \
+    obj/validationinterface.o \
     obj/version.o \
     obj/support/cleanse.o \
     obj/checkpoints.o \
