+++ src/makefile.osx
@@ -72,6 +72,7 @@ CFLAGS += -Wall -Wextra -Wformat -Wno-ignored-qualifiers -Wformat-security -Wno-
 OBJS= \
     obj/alert.o \
     obj/allocators.o \
+    obj/validationinterface.o \
     obj/version.o \
     obj/support/cleanse.o \
     obj/checkpoints.o \
